requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.19
langchain==0.2.6
langchain-openai==0.1.13
langchain-anthropic==0.1.15
langchain-community==0.2.6
redis==5.0.6
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
langchain-groq==0.1.6
slowapi==0.1.9
